Factor out client request adapter test from elicitations tests
